requirements.txt
Flask==2.3.3
openai==1.14.3
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.1
flask-cors==4.0.0
google.generativeai
gunicorn==21.2.0
Flask
Flask-Session
Flask-SQLAlchemy
Werkzeug
python-docx
fpdf
supabase
fpdf
python-docx
pillow
